Navigator: Enhancements, fixes and macOS support
- Fixed the totally broken file path detection. It wasn't resolving the
current path when files were opened with relative paths (such as "mpv
*.mkv" in a folder). So I rewrote everything to check for an absolute
path, and if not absolute then we prepend mpv's current working dir.
- Made the path detection resolve relative paths ("/foo/bar/../bar/"
becomes "/foo/bar/").
- Added final failsafe so that if the path detection came up with a
folder that doesn't actually exist, we fall back to starting the
navigation from mpv's current working dir instead.
- Rewrote the "isfolder" check in a more efficient, cross-platform way
so that it works on both Linux and macOS.
- Rewrote the folder scanner. Instead of wastefully using the overkill
"find" utility (along with non-compatible flags that don't work on
macOS), we now use "basename" instead and simply pass it a glob of all
files in the folder. It returns the basenames (filename without path)
for all files. That utility exists on both Linux and macOS, which means
that the script is now supported on macOS.
- Added a new function which clears the OSD, removes keybindings, and
restores the font size to normal. And made all "reset/close/cancel"
functions and timers use that instead.
- Fixed the broken "safe path" generator. It was replacing all sequences
of all kinds of whitespace with a single "\ " (a backslash and regular
space. That meant that it didn't support folders with non-space
whitespaces or more than a single space in a row. The new code solves
that by properly escaping every whitespace character one by one.
- Added useless macOS system paths to the default ignore list.
- Enhanced keybinding of the main key so that you can choose it in the
"settings" and can even set it to "nil" to allow you to bind it via
input.conf instead.
- Made it possible to rebind the "favorites" key via "settings".
- Simplified favorites configuration (there is no need to number keys).
- Added setting to choose how many OSD items to display per screen.
- Added automatic font size switching, so that you can use a smaller
font to see lots of files while the navigator is open and then return to
your normal OSD font size afterwards.
- Replaced all tabs with two spaces, because oh my god stop mixing tabs
in half the file and two spaces in the other half! ;-)